requirements.txt
fastapi
uvicorn
google-genai
yt-dlp
pydantic
